# models/category.go
package models

import (
	"time"

	"gorm.io/gorm"
)

type Category struct {
	ID          uint           `gorm:"primarykey" json:"id"`
	CreatedAt   time.Time      `json:"created_at"`
	UpdatedAt   time.Time      `json:"updated_at"`
	DeletedAt   gorm.DeletedAt `gorm:"index" json:"deleted_at,omitempty"`
	Name        string         `json:"name"`
	Description string         `json:"description"`

	PersonID uint   `json:"person_id"`
	Person   Person `gorm:"foreignKey:PersonID" json:"-"`
}