# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sipelan API",
        "version": "2.0.0",
        "description": "API for managing Categories, Transactions, and Dashboard Summary"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "summary": "List all categories",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedCategoryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new category",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "summary": "Get a category by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "summary": "Update a category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "summary": "List all transactions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by transaction type",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pemasukan",
                                "pengeluaran"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedTransactionResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new transaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "summary": "Get a transaction by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction found"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "put": {
                "summary": "Update a transaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction updated"
                    },
                    "404": {
                        "description": "Transaction not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a transaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction deleted"
                    }
                }
            }
        },
        "/summary": {
            "get": {
                "summary": "Get overall summary (total pemasukan, pengeluaran, balance)",
                "responses": {
                    "200": {
                        "description": "Summary retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OverallSummary"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/summary/monthly": {
            "get": {
                "summary": "Get monthly income and expense totals for chart",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Year to get data for (default: 2025)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly summary retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MonthlySummary"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "CategoryInput": {
                "type": "object",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "Transaction": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "description": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "pemasukan",
                            "pengeluaran"
                        ]
                    },
                    "attachment": {
                        "type": "string"
                    }
                }
            },
            "TransactionInput": {
                "type": "object",
                "required": [
                    "date",
                    "category_id",
                    "total",
                    "type"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "pemasukan",
                            "pengeluaran"
                        ]
                    },
                    "attachment": {
                        "type": "string"
                    }
                }
            },
            "OverallSummary": {
                "type": "object",
                "properties": {
                    "total_pemasukan": {
                        "type": "number"
                    },
                    "total_pengeluaran": {
                        "type": "number"
                    },
                    "balance": {
                        "type": "number"
                    }
                }
            },
            "MonthlySummary": {
                "type": "object",
                "properties": {
                    "month": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "pemasukan": {
                        "type": "number"
                    },
                    "pengeluaran": {
                        "type": "number"
                    }
                }
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "total_items": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    }
                }
            },
            "Response": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
            "PaginatedCategoryResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                }
            },
            "PaginatedTransactionResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Transaction"
                        }
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                }
            }
        }
    }
}